Fix typo in match orderings

+++ cranelift/codegen/src/isa/riscv64/lower/isle.rs
@@ -539,7 +539,7 @@ impl generated_code::Context for RV64IsleContext<'_, '_, MInst, Riscv64Backend> 
     }
 
     fn fcvt_smin_bound(&mut self, float: Type, int: Type, saturating: bool) -> u64 {
-        match (float, int) {
+        match (int, float) {
             // Saturating cases for larger integers are handled using the
             // `fcvt.{w,d}.{s,d}` instruction directly, that automatically
             // saturates up/down to the correct limit.
@@ -560,7 +560,7 @@ impl generated_code::Context for RV64IsleContext<'_, '_, MInst, Riscv64Backend> 
 
     fn fcvt_smax_bound(&mut self, float: Type, int: Type, saturating: bool) -> u64 {
         // NB: see `fcvt_smin_bound` for some more comments
-        match (float, int) {
+        match (int, float) {
             (I8, F32) if saturating => f32::from(i8::MAX).to_bits().into(),
             (I8, F64) if saturating => f64::from(i8::MAX).to_bits(),
             (I16, F32) if saturating => f32::from(i16::MAX).to_bits().into(),
@@ -574,7 +574,7 @@ impl generated_code::Context for RV64IsleContext<'_, '_, MInst, Riscv64Backend> 
 
     fn fcvt_umax_bound(&mut self, float: Type, int: Type, saturating: bool) -> u64 {
         // NB: see `fcvt_smin_bound` for some more comments
-        match (float, int) {
+        match (int, float) {
             (I8, F32) if saturating => f32::from(u8::MAX).to_bits().into(),
             (I8, F64) if saturating => f64::from(u8::MAX).to_bits(),
             (I16, F32) if saturating => f32::from(u16::MAX).to_bits().into(),
